Build ready PR (#7)
* New Interfaces Added

Part 1 of the new interfaces. `LinkedList` was the first class to implement `Insertable` interface, Next part will be to actually implement them to all data structures.

Also added a constructor for an empty pair.

* Arrays implement the new Insertable Interface

Re-implemented the `Array` and `Vector` to use the `Insertable` interface alongside the `AbstractArray` class. Updated the tests and other classes to match the new structure of the arrays.

* Added the `Maps` data structure

Re-implemented the `HashTable` data structure and made it in to a `HashMap` instead. No test yet for the class. Have to redo the `BinaryTree` tests.

* Refactored a couple classes

Improve the `Iterator`s for the following classes: `LinkedList`, `Array` and `Vector`.

Added JavaDoc for the `HashMap` class. Also added tests for the hash map.

Refactored `Pair#toArray()` and refactored the `PairTest`

Changed the scope of tests to `compile` so that you can now run tests without using maven. (Might change later)

* Full implementation of the `Pushable` interface

`Stack` and `Queue` implements the new `Pushable` interface. As such all other classes that utilize those classes needed to be updated. Also fixed the `Stack` class so that it conforms to the rule of LIFO.

Renamed the `Insertable#append` to `Insertable#add` to be more consistent with other implementations.

* Removed the debugger line in `StackTest`

* Removed the binary trees

Need to reimplement `BinaryTree` and `BinarySearchTree` as the overarching changes broke the classes.

* Added JavaDoc for the rest of the classes

Also fleshed out the pom.xml for exporting the .jar